Properly handle normalization in utf8-c8.
We can emit NFG synthetics into the result string, but only when the
result of encoding them back to utf8-c8 later will not result in the
byte sequence produced changing. Also, any normalization that changes
the codepoints in any way will result in synthetics being inserted to
ensure the round-trip.

This passes the utf8-c8 tests that we passed before. It also passes
encode.t without upsetting valgrind, and makes todo'd tests for
utf8-c8 pass also, which will fix an RT.