Do gen2 sweep before letting stolen thread go.
Sweeping gen2 involves putting things onto a free list. Nursery sweeps
are safe to do concurrently with the mutator running since there is no
free list, just a now-unused fromspace that another thread is free to
consider independently. However, gen2 sweeps manipulate a free list
that is used to allocate, so the mutator runing at the same time would
race and potentially corrupt this free list.

To fix this, do the gen2 sweep before we mark the thread as unstolen.
This actually does it whether the thread was stolen or participating,
so we could optimize this fix a bit more for the participatory case
when the concurrent operation is safe.