Speed up collapse_strands 2-4x by full rewrite targeting SIMD optims
collapse_strands and certain conditions of substr rely on
iterate_gi_into_string() for flattening a specified section of a string.
This has been a performance issue since forever, and though I made
changes to make it faster, it still had issues. These new changes are a
full rewrite of that function and a new function which checks if a
string for the specified length can fit into 8 or 32 bits.

Both these functions have been rewritten to take advantage of compiler
vectorization which will utilize instructions such as SSE* or AVX*.

In string_can_be_8bit() we use multiple bitwise AND operations to check
the 32 bit integers while ORing the result with `val`.
Care must be taken so that any variables altered by the loop are not used
outside of the loop. Even more critical, there cannot be  any branching or
function calls. For example: `i` is not used outside the loop, while `val`
is allowed as biwise OR works with vectorization well.

To support both of these changes a new grapheme iterator function has
been added called MVM_string_gi_next_strand() which moves to the next
strand. This allows us to process each strand at a time without having
to process each grapheme at a time. This was needed to either memcpy()
or use vectorizable loops to copy the graphemes based on strand bit
type.