Randomly seed a hash secret to mitigate ease of DOS attack
Without a random hash seed it is easy for an attacker to generate
strings which will result in the same hash. This devolves to O(n**2)
time for the hash insertion and is increased even more by the fact that
a string must be compared with all strings in the bucket until the right
string is found.

This attack is done by creating a function that essentially is our
hashing function backward. We hash our target string, `t`. We then
use random 3 character sequences (in our case graphemes) and plug them
into our backward hashing function along with the hash for our target
`t`. The backward hash and the random character sequence are stored in
the dictionary and the process is repeated until we have a very large
number of backward hash's and random 3 grapheme prefixes.

We can then use this dictionary to construct successively longer strings
(or short if we so desire) which are the same hash as our target string
`t`.

This has been fixed in most programming languages (Python, Ruby, Perl),
and several CVE's have been issued over the years for this exploit.

It may also be a good idea to later implement a stronger hashing
function. Many languages are now using SipHash which is meant to protect
against an attacker discovering a hash secret remotely. This change does
increase the ease of this attack greatly and makes the hash secret
unpredictable.

Randomness source:
We prefer function calls rather than reading from /dev/urandom. Reasons
include: not having to open a file descriptor and that /dev/urandom may
not exist if we are in a chroot. Since we don't want to stop startup of
MoarVM, if we have to fallback to /dev/urandom and it doesn't exist (we
are in a chroot) we continue starting MoarVM normally.
Linux, FreeBSD, OpenBSD and MacOS all use system provided random calls
to get the data rather than having to open /dev/urandom. All these OS's
guarantee these to be non-blocking, though MacOS's documentation does
not comment on it.
Whether the calls block is primarily a concern during very early boot
which is why Python 3 makes sure to use non-blocking calls when their
hash secret is seeded an interpreter start up.

This change was tested on Linux both pre-addition of `getrandom()` and
after, as well as Solaris, FreeBSD and OpenBSD. CI testing for Windows
and MacOS look good. All Unix are supported due to /dev/urandom fallback
and Windows is supported to version 95 with the API we use.