Get rid of lots of mallocs in coerce_sI
This avoids mallocing in MVM_bigint_from_str if we just end up storing
the value as a smallint. It also avoids mallocing in MVM_coerce_sI when
creating a cstring version of the input string. We can shortcut and use
the faster grapheme iterator since decode will fail with synthetics
anyway.